backend/pyproject.toml
[project]
name = "rarefactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "greenlet>=3.2.4",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.0.1",
    "sqlmodel>=0.0.27",
]
